pyproject.toml
[project]
name = "postgres-data-generator"
version = "0.1.0"
requires-python = ">=3.11,<3.12"

[tool.poetry.dependencies]
SQLAlchemy = "2.0.30"
psycopg2-binary = "2.9.10"
Faker = "25.2.0"
PyYAML = "6.0.2"
